places/admin.py
from adminsortable2.admin import SortableAdminBase, SortableTabularInline
from django.contrib import admin
from django.utils.html import format_html

from .models import Place, PlaceImage


class PlaceImageInline(SortableTabularInline):
    model = PlaceImage
    extra = 1
    readonly_fields = ('place_images',)

    def place_images(self, place_image):
        return format_html('<img src="{}" height="{}"', place_image.image.url, '200px')


@admin.register(Place)
class PlaceAdmin(SortableAdminBase, admin.ModelAdmin):
    inlines = [PlaceImageInline]
